+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from "./Session";
 export * from "./createSession";
+export * from "./createResponse";
 
 export * from "./Channel";
 export * from "./createChannel";